Properly support multi-selected cells for delete (including undo/redo)
The delete cell action uses the `{ multiSelect: true }` option to handle multi-select scenarios. This causes `cell.delete()` to be called once for each selected cell, creating separate undo operations for each cell which is not correct for a multi-select scenario. When cells are deleted via multi-select, the deletes should happen as one atomic transaction and undo should bring back all deleted cells in the multi-select.

To fix these, we are using the `deleteCells` function instead of the `deleteCell` function which internally handled multiple cells and the undo/redo logic for all cells the operation should work.

I have also updated the action to extend NotebookAction2 to clearly communicate that this action relies on the notebook instance and is not run per cell.