Update mulit-select supported  actions to use NotebookAction2 class
Actions that rely on a function which internally handled multi-select scenarios and undo/redo history should not use the CellAction2 class. The CellAction2 class for multi-select scenarios applies the action to each cell individually and tracks each operation in the undo/redo history as a separate transaction. This is confusing for users because the expectation is that all cells are part of a single transaction as far as undo/redo is concerned.

For example: When cutting 3 cells together, the undo operation should bring all 3 cells back. Prior to this change, undo would only bring the last cell back and the user would have to undo the operation per cell.