Add assistant reset state and generate diagnostics commands (#10630)
Addresses #10344 and #10346

Adds two new commands:

- Positron Assistant: Collect Diagnostics
- Positron Assistant: Reset State

The Reset State command brings up a modal with some details of what is
being reset and giving the option to generate a diagnostics report prior
to clearing the state. If generated, the diagnostics report is saved to
the user's data directory and and is open when the window reloads.

<img width="668" height="301" alt="image"
src="https://github.com/user-attachments/assets/d388212e-68b3-46e2-bf4c-44ba7e9038c4"
/>

Resetting the state performs the following tasks:

- Optionally saves the diagnostics report
- Signs out of providers
- Deletes all global state keys that start with `positron.assistant`
- Deletes any API Keys associated with models
- Clears chat history with the following commands:
 - `workbench.action.chat.clearHistory`
 - `workbench.action.chat.clearInputHistory`
- Reloads the window

Progress is provided via notifications about the state reset.

The diagnostics report is a markdown file generated in an editor window.
It includes the following:

- A privacy notice informing the user what information is included in
the report and instruction to review before sharing
- Version information for Positron Assistant and Copilot extensions,
Positron, Code OSS
- Positron Assistant settings set to non-default values
- Providers that the user has signed into
- Available Models (including whether an API key is configured, but not
including the key)
- Information about the active chat session
- The last 500 Positron Assistant logs (available from "Positron
Assistant" in the output pane)

To retrieve the Positron Assistant logs I created a wrapper class around
`LogOutputChannel` that uses a circular buffer to track the most recent
log entries. We could modify the copilot extension to use this class as
well if we think including those logs on the report is beneficial.
 
### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Assistant: add command "Positron Assistant: Reset State" to reset
assistant state https://github.com/posit-dev/positron/issues/10346
- Assistant: add command "Positron Assistant: Collect Diagnostics" to
generate diagnostics report
- 
#### Bug Fixes

- N/A


### QA Notes
The only assistant state remaining after a reset should be the
configured user settings.
@:assistant