Merge upstream 1.106.0 into Positron (#10637)
Updates Positron with changes from VS Code October 2025:

https://code.visualstudio.com/updates/v1_106

Addresses https://github.com/posit-dev/positron/issues/10349. 

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Update Positron to upstream [Code OSS
1.106.0](https://code.visualstudio.com/updates/v1_106) (#10349).

---------

Co-authored-by: positron-bot[bot] <173392469+positron-bot[bot]@users.noreply.github.com>
Co-authored-by: Marie Idleman <marie.idleman@posit.co>