Add support for empty notebook to context key
This fix updates the POSITRON_NOTEBOOK_EDITOR_CONTAINER_FOCUSED key to not be false when there are no cells in a notebook.

This ensures the paste action (and any other action that uses the POSITRON_NOTEBOOK_EDITOR_CONTAINER_FOCUSED context key) to only fire when the notebook editor has focus!

We can't rely on the POSITRON_NOTEBOOK_CELL_EDITOR_FOCUSED context key because it is false when there are no cells in the notebook and we can't rely on the activeEditor context keys because it is always true and causes the action to fire EVERY time a user presses v in an input.