catalog-explorer: Surface the Snowflake user and role in the tooltip.
Previously we showed the user and role from the `connections.toml` file
in the tooltip for a Snowflake catalog, but this information is often
missing or misleading (depending on the authentication method in use),
so Chris Ostrouchov suggested we retrieve it from the connection itself.

The existing `CatalogProvider` interface already has a mechanism for
this kind of dynamic tooltip (via the `getDetails()` method), so this
commit merely wires it up to an appropriate Snowflake SQL query.

Note that the query only runs once when the user hovers over the catalog
provider for the first time, so this doesn't introduce additional
overhead or increase the initial loading time the tree view.

In order to make things look a little nicer this also expands the
catalog API a bit to allow for using Markdown in tooltips.

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>