Assistant: revisit approval workflow in `editFile` (#10672)
Addresses #10671.

With this PR:


https://github.com/user-attachments/assets/be338986-4ea9-4027-983d-31867d0ad8d5

On tool request, rather than returning after applying edits, the tool
waits for Keep/Undo. On Keep, the tool returns as before. On Undo, a
tool result "The user rejected the edit and might provide feedback." is
appended to the conversation history but an interrupt that fires
immediately after prevents the model from then generating a response /
another attempt and allows the user to provide feedback. At that point,
the new placeholder text "Provide feedback on the proposed edit..."
queues the user to provide feedback.

Notably, once the user clicks "Undo", they can't see the change that
they're providing feedback on, which is a bummer. (I tried to implement
this but ended up in over my head.🫠)

For multi-part Edits, the Keep/Undos for individual changes still work
as they did before.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->