Add a button to stop a terminal process from the Viewer pane (#10625)
Addresses https://github.com/posit-dev/positron/issues/7747

@:apps
@:viewer

This PR extends the UI comm for previewing a URL to optionally keep
track of the "source" of where that URL may be coming from, for example,
the terminal that may be running that process. If we have such a process
ID, then we show a new button the Viewer pane UI that can be used to
stop the process. It's the same button that we use on the console to
stop running code there.

<img width="1532" height="1055" alt="Screenshot 2025-11-21 at 4 14
36 PM"
src="https://github.com/user-attachments/assets/7cdef6aa-d3e6-4331-adbb-1242af8d1eaa"
/>

I tried a few hacky things that might allow us to show such a button
without truly tracking the source, but it did NOT result in good
behavior. We can do PRs to the Shiny extension and the Quarto extension
so they use this new API and can also show buttons that will stop their
processes as well.

We also can extend this to process that come from the language runtimes,
but I did not do that here. Do people want *two* buttons in such a case?
The button that would be running in the console and the button in the
viewer? A little weird maybe? I say let's wait and see what folks think.

### Release Notes

#### New Features

- Added new button to interrupt a running app (such as Streamlit or
FastAPI) from the Viewer pane

#### Bug Fixes

- N/A


### QA Notes

If you have an app supported by the builtin `positron-run-app`
extension, when you press the "play" button, you should see a new button
pop up on the Viewer pane.

- If you click the button, it should stop the app from running (similar
to if you stopped it from the terminal)
- Having another terminal focused or active should not interfere with
the button behavior; the button should act like it is connected exactly
to the process running the app

Be aware that we do not have clearing set up in the Viewer pane, i.e.
https://github.com/posit-dev/positron/issues/6827,
https://github.com/posit-dev/positron/issues/9323,
https://github.com/posit-dev/positron/issues/5344. Adding the button
here might make that feel more urgent.