notebook multiselect revamp for actions  (#10678)
Addresses #10116 and #10397

This PR revamps a good portion of our notebook cell actions to support
multi-selections and ensure that undo/redo work as expected.

Updated actions that now support multi-select properly with undo/redo:
- cut
- delete
- copy
- paste
- move cell up
- move cell down

I also fixed #10482 which fixes the "Run Below Cells" action to also
include the current cell in its execution.

**Run Below Cells - BEFORE**


https://github.com/user-attachments/assets/ed45c886-46a7-470d-a416-9fbe38329f0e

**Run Below Cells - AFTER**


https://github.com/user-attachments/assets/2d1f91d6-855d-4974-9bac-f4929210f3bc

**Move Cells with Undo- AFTER**


https://github.com/user-attachments/assets/ec65dfff-77fd-4c87-b7e3-df0ba20fac41

**Cut/Copy/Paste/Delete with Undo- AFTER**


https://github.com/user-attachments/assets/e82354fc-2832-41bb-8852-4d12a5064737


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

@:positron-notebooks

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->

---------

Co-authored-by: Marie Idleman <marie.idleman@posit.co>