python LSP: disable Jedi language server features redundant with Pyrefly (#10770)
Addresses https://github.com/posit-dev/positron/issues/10305. Updates
the language server code so it disables the capabilities that are
redundant with Pyrefly, except completions. (There will still be
duplicate completions.)

This won't be super maintainable in the long run, so I'll revisit this
file next week and really slim it down, hopefully.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->
@:notebooks @:web @:win @:help

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->
Language servers have a lot of surface area so there's a lot to verify
here, sorry. We definitely want to make sure there's no more duplicate
contributions for hover and code outlines. Completions for DataFrame
columns, environment variables, and magics should still work, though
completions may still be doubled in some cases. The help pane should
work.