Auto refresh models after adding a new provider (#10799)
Addresses #10793 

The 1.106.0 upstream merge brought in a change intended to improve
efficiency that prevented model registration when adding a provider that
wasn't already configured by a user. This prevented the model list from
refreshing when a user signed into a new provider.

This change resolves that issue. I left the conditional statement
commented out in hopes that it would make things easier with the next
upstream merge, but I'm open to other ways of handling this.

### QA Notes
Repro notes detailed on original issue 
@:assistant