Make 'no language model' error message more generic (#10804)
Addresses #9823 

When the original issue was filed, Assistant would fail silently when
the user sent a message without a model configured. In doing my own
testing, Assistant was providing a helpful message, but it called out
'Anthropic' as the only potential provider. This PR updates the wording
of that error to be more generic.


### QA Notes

@:assistant