Implement soft-fail approach for flaky positron-r tests on Windows (#10771)
This PR is about two categories of intermittent failures we observed
with the extension tests of `positron-r` on Windows CI.

- The most recent concerns the ArkComm tests, context at
https://positpbc.slack.com/archives/C07FR1JNZNJ/p1763576791511939?thread_ts=1763575446.108129&cid=C07FR1JNZNJ

- A couple months ago we had to deal with debugger test failures at
<https://positpbc.slack.com/archives/C07FR1JNZNJ/p1761068770237129?thread_ts=1761067781.072859&cid=C07FR1JNZNJ>.
The PR that disabled these tests (on all platforms rather than just
Windows) is at <https://github.com/posit-dev/positron/pull/10190>.


The goal of the PR is to implement an informative soft-fail approach:

- Keep running all the tests on all platforms instead of disabling tests
observed to be flaky. I would like to maintain the test coverage
established in https://github.com/posit-dev/positron/pull/8816.

- Provide full logs when a test fails so that we get better context on
the failures.

- Prevent these flaky tests from being a source of distraction and
friction for the QA and dev teams, while we figure them out.


Approach:

- All tests are reenabled.

- The extension tests no longer cause the whole e2e-windows job to fail.
Instead we upload logs and notify a new Slack channel
`#positron-rstats-test-results`.

- To get proper context on tests in the logs, the `sessionName` argument
of `startLanguageRuntime()`, which was unused, is now used to update the dynamic session name.

- The `startR()` utility of positron-r test now looks up the currently
running test and passes that as `sessionName` to
`startLanguageRuntime()`. If a session is created for a whole suite, a
new optional argument can be supplied to provide the suite name (there's
no way to look it up from `startR()`).

- The kernel supervisor and the lsp manager in positron-r now include
the `sessionName` in the opening line of logs when a session is started.
This allows the test titles to be inserted in the logs whenever a
session start, and provides "fences" between test runs (unless sessions
are run in parallel but that's normally not the case). Since that
changes `positron.d.ts`, I've asked you for a review @jmcphers.

- The PR also makes sure all relevant output channels are marked as logs
so that they get exported as logs. The Ark LSP channel was not marked as
logs.


### QA Notes

Since `sessionName` might change some details captured by tests, I ran
the full suite at
<https://github.com/posit-dev/positron/actions/runs/19647373986>. The
failures seem unrelated.

Here is how a failing run of e2e-windows looks like:
<https://github.com/posit-dev/positron/actions/runs/19650448090>. That
run did succeed as a whole but the logs were uploaded as artifacts and a
notification was sent to the new slack channel.

And here is a success run of e2e-windows:
<https://github.com/posit-dev/positron/actions/runs/19650462016>