optimize excluded results check with early exit
I noticed that the second, unrestricted search in case of sparse result could take a while if there were many results, so I opted to stop after _any_ additional results were found in the second-search case and then just note that _something_ was found. This cuts down on the latency quite a bit, and the model seems to still take the hint that it should keep looking.