Get rid of big "Run and Debug" button for R extension (#10769)
Addresses https://github.com/posit-dev/positron/issues/3681

This is a little heavy handed (I would have liked to find a solution
here that involves a less dramatic change) but this area of upstream
doesn't seem to involve a lot of churn, so I think it's likely useful
for us to make this type of adjustment. Quarto files are another type
where this whole pane is quite confusing, and we can use a similar
approach in that extension (set `supportsUILaunch` and contribute a
`viewsWelcome` button or something).

With this PR, you now see this in the "Run and Debug" pane for `.R`
files:

<img width="1418" height="951" alt="Screenshot 2025-11-24 at 9 45 20 PM"
src="https://github.com/user-attachments/assets/1c500c41-72b4-4d26-bce9-2af84de5f716"
/>

You should still see the normal content for a `.py` file or `.qmd` file
or another kind of file.

<img width="1418" height="951" alt="Screenshot 2025-11-24 at 9 45 30 PM"
src="https://github.com/user-attachments/assets/9e056ed8-394d-4d7e-bdc3-baf79c4e1955"
/>


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- R: added a link to the "Run and Debug" pane with info on debugging R
code.

#### Bug Fixes

- N/A


### QA Notes

- You should see new, less confusing content in the "Run and Debug" pane
for `.R` files.
- You should see no change to this pane for `.py`, `.qmd`, or other
types of files.

---------

Signed-off-by: Julia Silge <julia.silge@gmail.com>
Co-authored-by: Jonathan <jonathan@rstudio.com>
Co-authored-by: Lionel Henry <lionel.hry@proton.me>