Fix cleanup of R sessions (#10933)
Progress towards #10121 

Logs involving the session manager are a bit confusing because sessions
are never cleaned up, and so the manager keeps trying to deactivate LSP
in sessions that no longer exist. This also causes it hold onto
`RSession` objects, preventing garbage collection.

To fix this:

- `RSession` gains a `register()` method for external resources that
should be cleaned up when the session is disposed of.
- The manager's event handler is now registered in `RSession`, instead
of the singleton's own disposables.
- The manager no longer maintains a list of session and instead queries
the main thread for active sessions. Because of this process crossing a
bunch of methods are now async.

<!--
  Describe briefly what problem this pull request resolves, or what
  new feature it introduces. Include screenshots of any new or altered
  UI. Link to any GitHub issues but avoid "magic" keywords that will
  automatically close the issue. If there are any details about your
  approach that are unintuitive or you want to draw attention to, please
  describe them here.
-->


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->

@:ark