Add experimental Dev Container support to Positron (#10884)
This change adds early, experimental support for Dev Containers to
Positron.

<img width="1226" height="860" alt="image"
src="https://github.com/user-attachments/assets/e0ca3360-12c9-47da-93c6-cade37db47a4"
/>

Because the feature is experimental, it is currently behind a feature
flag:

<img width="578" height="136" alt="image"
src="https://github.com/user-attachments/assets/fd2495e9-dc88-436a-a0c7-035bf75623fa"
/>

The goal of this change is to enable the most common dev container
workflows that center around building (or rebuilding) a container and
opening the workspace inside the container. However, note that in this
early version, many features are not implemented or are incomplete.
Specifically:

- musl-based Linux distributions (such as Alpine) are not supported
(#10468); the container must be running a version of Linux already
supported by Positron
- Only single-container configurations are supported (no `docker
compose`)
- No tools or templates are supplied for creating or authoring
`devcontainer.json` files
- Only one container is supported per project
- Docker volumes are not supported, just regular mounts
- In the MRU, dev containers show up with the remote rather than local
path (e.g. `/workspaces/foo`)
- Container management features (see/attach to running containers) are
not available when inside a container

For compatibility, the new extension uses largely the same command names
and settings as the official VS Code Dev Containers extension.

It also includes a lightly modified version of the [Dev Containers
CLI](https://github.com/devcontainers/cli) reference implementation. The
implementation is used to derive the correct container management
commands, which are (generally) executed in Positron terminals for
visibility.

Addresses #4691. 



### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Experimental support for Dev Containers (#4691)

#### Bug Fixes

- N/A


### QA Notes

Like Remote SSH, this feature is difficult to test outside release
builds since it needs to download a matching server component into the
container. And -- also like remote SSH -- there's a setting to change
the download URL. If you need to test a build that doesn't have a
released Positron version to download, you can change this setting to
the URL of the closest matching build.

This feature should be tested across all three operating systems since
the Docker integration is somewhat OS-specific.

---------

Co-authored-by: positron-bot[bot] <173392469+positron-bot[bot]@users.noreply.github.com>