Update inno_updater.exe branding (#11005)
Address #9809 

Updates the icon and metadata on the executable. The strings within the
executable have been updated to say Positron instead of VS Code. The log
file name has been changed to `positron-inno-updater-1765224032.log`
(the numbers being the timestamp when updating).

The logs will mention Positron now when referring to the application.

The changes to the inno_updater source was a search and replace of
`VSCode|Visual Studio Code|VS Code` to `Positron`

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- Fix branding on process that finalizes the Positron update on Windows


### QA Notes

The log file may be found in `AppData\Local\Temp\1` where `1` may be a
different number. The auto-update mechanism can be used to test this as
the new `inno_updater.exe` is installed before the restart to finalize
the update.