Delegate Console copy operation to the document
The Console renders input and output using HTML, which sometimes will
replace spaces with non-breaking spaces. When copying the string and
subsequently pasting it back into the Console, the non-breaking space
becomes an invalid character. Delegating to the document to perform the
copy command results in a string with normal spaces.

The built-in copy command will convert the string, which might contain
uncommon characters, back into what we expect.

VS Code's built-in terminal executes the same copy command when working
with HTML contents, so there is prior art to using this method.

See: `xtermTerminal.copySelection()`

Another solution would be to handle the non-breaking space (0x00A0) on
its own like so, but it does not seem like the best generalized
solution with the assumption there are other characters that might be
need to be replaced (e.g. maybe quotes?)

```ts
string.replace(/\u00A0/g, ' ')
```